% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignStage.R
\name{AssignStage}
\alias{AssignStage}
\title{Assign worm stages}
\usage{
AssignStage(Channel)
}
\arguments{
\item{Channel}{Channel data}
}
\value{
A list with 6 stages of worms
}
\description{
The function takes in the channel data and sorts the worms into six different stages (L1, L2,L3, L4,Young Adult, Adult)
depending on their size.
}
