% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTrainingSetIDs.R
\name{CreateTrainingSetIDs}
\alias{CreateTrainingSetIDs}
\title{Read the categorized worm IDs file}
\usage{
CreateTrainingSetIDs(StageChannelData, IndexDirectory)
}
\arguments{
\item{StageChannelData}{Channel Data.}

\item{IndexDirectory}{Path to the GoodIDs file.}
}
\value{
A list with Good worms IDs and Bad worm IDs.
}
\description{
The function takes in the channel data and the directory of the GoodIDs file. The GoodIDs
is an output of the app WormProfiler which includes the worm IDs and whether the worm is a good worm
or a bad worm. The output is a list containing the IDs of the good worms that are in the channel data, and the IDs
of the bad worms.
}
