% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterChannel.R
\name{FilterChannel}
\alias{FilterChannel}
\title{Filter channel data}
\usage{
FilterChannel(ch, MaxTOF, MinLength, MaxLength)
}
\arguments{
\item{ch}{Channel data}

\item{MaxTOF}{Maximum time of flight}

\item{MinLength}{Minimum length}

\item{MaxLength}{Maximum length}
}
\value{
Returns the index number of the filtered objects
}
\description{
This function takes in the channel data and filters it according to the input parameters. The parameters are:
Max Time of flight, Minimum length, and maximum length.
}
