% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPrediction.R
\name{GetPrediction}
\alias{GetPrediction}
\title{Get prediction from a training set}
\usage{
GetPrediction(SetList)
}
\arguments{
\item{SetListr}{List input that has training data set and prediction data set.}
}
\value{
Prediction output
}
\description{
The function takes in a list input that has the training set data and the prediction set data. The function then
runs the svm function to get a prediction and returns it to the user.
}
