% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAccuracy.R
\name{PlotAccuracy}
\alias{PlotAccuracy}
\title{Plots the accuracy}
\usage{
PlotAccuracy(Accuracy, SetNumber, BadWormSampleSize)
}
\arguments{
\item{Accuracy}{The accuracy data set from the bootstrap for loop}

\item{SetNumber}{The vector containing the different numbers of good worms in the training set}

\item{BadWormSampleSize}{How many bad worms in the training set}
}
\value{
Returns the accuracy plot of the model
}
\description{
The function takes in the accuracy vector and the Set number and the bad worm sample size, and plots
the accuracy of the model against different number of good worms in the training set.
}
