% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessData.R
\name{PreProcessData}
\alias{PreProcessData}
\title{Pre-process data}
\usage{
PreProcessData(ChannelData, MinTOF, MaxTOF, MaxAmp, ChannelNumber)
}
\arguments{
\item{ChannelData}{Channel list output from the ReadChannel function}

\item{MinTOF}{Minimum TOF accepted for each profile}

\item{MaxTOF}{Maximum TOF accepted for each profile}

\item{MaxAmp}{Maximum amplitude accepted for each profile}

\item{ChannelNumber}{Which channel to perform the pre-processing on. 1 is the optical density channel, 2, 3 and 4 are the green, yellow and red fluorescence channels respectively.}
}
\value{
Pre-processed channel data ready to train SVM model, to validate an SVM model and/or to be classified from an SVM model.
}
\description{
The function takes in the selected channel data and performs multiple pre-processing steps.
\enumerate{
\item Removing profiles outside of the provided time of flight range and profiles with optical density that exceed the provided
maximum amplitude input
\item Trimming trailing zeros for each profile
\item Applying Piecewise Aggregate Approximation (PAA function from the jmotif library) to uniform the length of all profiles to the length of the minimum time of flight minus one.
\item Performing Z score normalization on the Y axis using the scale function.
}
}
