% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadChannel.R
\name{ReadChannel}
\alias{ReadChannel}
\title{Read Channel Data}
\usage{
ReadChannel(Ch0Directory, Ch1Directory, Ch2Directory, Ch3Directory)
}
\arguments{
\item{Ch0Directory}{Path to the channel 0 data.}

\item{Ch1Directory}{Path to the channel 1 data.}

\item{Ch2Directory}{Path to the channel 2 data.}

\item{Ch3Directory}{Path to the channel 3 data.}
}
\value{
List with the four reshaped data sets.
}
\description{
The function takes in the four channel data directories and loads them up. The returned output
is reshaped to a compatible form with the downstream functions.
}
