% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunClassification.R
\name{RunClassification}
\alias{RunClassification}
\title{Classifies input data using the input model}
\usage{
RunClassification(
  DataDirectory,
  ModelDirectory,
  MaxAmp,
  MinLength,
  MaxLength,
  TypeOfData = "FirstChannel"
)
}
\arguments{
\item{DataDirectory}{Directory of the COPAS output file.}

\item{ModelDirectory}{Directory of the model to be used, ends in .R}

\item{MaxAmp}{Maximum optical density, anything bigger will be removed.}

\item{MinLength}{Minimal object length, anything smaller will be removed.}

\item{MaxLength}{Maximum object length, anything bigger will be removed.}

\item{TypeOfData}{Type of COPAS output file. Either the first channel (optical density) file or the full unsplit channel file.}
}
\value{
Returns annotated IDs after classification
}
\description{
The function takes in the input data and runs classification using the input model.
}
