% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlot.R
\name{SummaryPlot}
\alias{SummaryPlot}
\title{Plots the provided Worm sorter output file}
\usage{
SummaryPlot(FileDirectory, Name, FluorescenceChannel, IDsFile)
}
\arguments{
\item{FileDirectory}{Directory to the worm sorter summary output file.}

\item{Name}{Name of the strain that will be used in the plots.}

\item{FluorescenceChannel}{Which channel to plot: G is green, R is red, and Y is yellow.}

\item{IDsFile}{Optional input: Directory to the annotated IDs file to remove the specified bad worms from the data set.}
}
\value{
Returns 6 plots in  a list.
}
\description{
The function takes in the data set directory, the name of the strain and plots the desired fluorescence channel.
The IDsFile parameter is optional. If the user provides it, the function will remove the bad worms from the
data set before plotting. The function returns five plots: 1- Worms with TOF from 50-75 2- Worms with TOF 75-150
3- Worms with TOF 150-225 4- Worms with TOF 225-500 5- Worms with TOF from 500-800 6- Box plot of the five TOF groups
}
