% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlots.R
\name{SummaryPlots}
\alias{SummaryPlots}
\title{Plots the provided worm sorter output files into a summary boxplot}
\usage{
SummaryPlots(
  FileDirectories,
  Names,
  FluorescenceChannel,
  Classify = "NA",
  Measure = "I",
  Scale = "Normal",
  TypeOfData = "Summary",
  MinMaxRange = c(50, 800),
  ModelDirectory = ""
)
}
\arguments{
\item{FileDirectories}{Directories to the worm sorter summary output file. In the format of <- c('Directory1','Directory2') and so on.}

\item{Names}{Names of the strains that will be used in the plots. In the format of <- c('Name1','Name2') and so on.}

\item{FluorescenceChannel}{Which channel to plot: G is green, R is red, and Y is yellow.}
}
\value{
Returns a summary boxplot of the fluorescence.
}
\description{
The function takes in the data set directories, the names of the strains and plots the desired fluorescence channel.
The boxplot provides a summary of the amplitude of the fluorescence of each strain.
}
