% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPrediction.R
\name{GetPrediction}
\alias{GetPrediction}
\title{Get prediction from a training set}
\usage{
GetPrediction(SetList)
}
\arguments{
\item{SetListr}{List input that has training data set and prediction data set.}
}
\value{
List containing the prediction and the model
}
\description{
The function takes in a list input that has the training set data and the prediction set data. The function then
runs the svm function to create the model and then runs the model against the prediction data. The function outputs
both the prediction of the model, and the model itself.
}
