% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlot.R
\name{SummaryPlot}
\alias{SummaryPlot}
\title{Plots the provided worm sorter output file (x Axis is Time of Flight and Y Axis is the amplitude of fluorescence)}
\usage{
SummaryPlot(
  FileDirectory,
  Name,
  FluorescenceChannel,
  IDsFile,
  Ranges = c(50, 75, 150, 225, 500, 800)
)
}
\arguments{
\item{FileDirectory}{Directory to the worm sorter summary output file.}

\item{Name}{Name of the strain that will be used in the plots.}

\item{FluorescenceChannel}{Which channel to plot: G is green, R is red, and Y is yellow.}

\item{IDsFile}{Optional input: Directory to csv file containing IDs of objects to be removed from data set before plotting.}

\item{Ranges}{The TOF ranges to group the objects with. Default is 50,75,150,225,500,800.}
}
\value{
Returns a summary boxplot of the fluorescence.
}
\description{
The function takes in the data set directory, the name of the strain and plots the desired fluorescence channel.
The function separates the objects in the data set based on their Time of Flight (TOF) and the Ranges input (the default is 50,75,150,225,500,800)
and returns 5 plots of the objects between those ranges. The user can also provide specific ranges instead.
The function also returns a sixth boxplot containing a summary of the grouped objects.
The user can also remove specific objects by providing the optional input: IDsFile which is a csv file containing the IDs of the objects
the user wants to remove from the data set.
}
